fix comment repo

+++ src/main/java/com/restbox/service/api/FetchTotalAmountBlogService.java
@@ -1,6 +1,6 @@
 package com.restbox.service.api;
 
 public interface FetchTotalAmountBlogService {
-    int fetchTotalAmountBlogByUsername(String  username);
-    int fetchTotalAmountBlog();
+    long fetchTotalAmountBlogByUsername(String  username);
+    long fetchTotalAmountBlog();
 }
